reduced the duplication in finding package name
