+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/generation/llm/Utils.kt
@@ -7,9 +7,42 @@ import org.jetbrains.research.testspark.core.progress.CustomProgressIndicator
 import org.jetbrains.research.testspark.core.test.Language
 import org.jetbrains.research.testspark.core.test.TestsAssembler
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
+import org.jetbrains.research.testspark.core.utils.javaPackagePattern
+import org.jetbrains.research.testspark.core.utils.kotlinPackagePattern
 import java.util.Locale
 
 // TODO: find a better place for the below functions
+
+/**
+ * Retrieves the package declaration from the given test suite code.
+ *
+ * @param testSuiteCode The generated code of the test suite.
+ * @return The package declaration extracted from the test suite code, or an empty string if no package declaration was found.
+ */
+// get package from a generated code
+fun getPackageFromTestSuiteCode(testSuiteCode: String?, language: Language): String {
+    testSuiteCode ?: return ""
+    return when (language) {
+        Language.Kotlin -> kotlinPackagePattern.find(testSuiteCode)?.groups?.get(1)?.value.orEmpty()
+        Language.Java -> javaPackagePattern.find(testSuiteCode)?.groups?.get(1)?.value.orEmpty()
+    }
+}
+
+/**
+ * Retrieves the imports code from a given test suite code.
+ *
+ * @param testSuiteCode The test suite code from which to extract the imports code. If null, an empty string is returned.
+ * @param classFQN The fully qualified name of the class to be excluded from the imports code. It will not be included in the result.
+ * @return The imports code extracted from the test suite code. If no imports are found or the result is empty after filtering, an empty string is returned.
+ */
+fun getImportsCodeFromTestSuiteCode(testSuiteCode: String?, classFQN: String): MutableSet<String> {
+    testSuiteCode ?: return mutableSetOf()
+    return testSuiteCode.replace("\r\n", "\n").split("\n").asSequence()
+        .filter { it.contains("^import".toRegex()) }
+        .filterNot { it.contains("evosuite".toRegex()) }
+        .filterNot { it.contains("RunWith".toRegex()) }
+        .filterNot { it.contains(classFQN.toRegex()) }.toMutableSet()
+}
 
 /**
  * Returns the generated class name for a given test case.
@@ -50,15 +83,7 @@ fun executeTestCaseModificationRequest(
     // Update Token information
     val prompt = "For this test:\n ```\n $testCase\n ```\nPerform the following task: $task"
 
-    var packageName = ""
-    testCase.split("\n")[0].let {
-        if (it.startsWith("package")) {
-            packageName = it
-                .removePrefix("package ")
-                .removeSuffix(";")
-                .trim()
-        }
-    }
+    val packageName = getPackageFromTestSuiteCode(testCase, language)
 
     val response = requestManager.request(
         language,
