+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/java/JavaJUnitTestSuiteParser.kt
@@ -1,12 +1,13 @@
 package org.jetbrains.research.testspark.core.test.java
 
 import org.jetbrains.research.testspark.core.data.JUnitVersion
+import org.jetbrains.research.testspark.core.generation.llm.getPackageFromTestSuiteCode
+import org.jetbrains.research.testspark.core.test.Language
 import org.jetbrains.research.testspark.core.test.TestBodyPrinter
 import org.jetbrains.research.testspark.core.test.TestSuiteParser
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.core.test.strategies.JUnitTestSuiteParserStrategy
 import org.jetbrains.research.testspark.core.utils.javaImportPattern
-import org.jetbrains.research.testspark.core.utils.javaPackagePattern
 
 class JavaJUnitTestSuiteParser(
     private var packageName: String,
@@ -15,7 +16,7 @@ class JavaJUnitTestSuiteParser(
 ) : TestSuiteParser {
     override fun parseTestSuite(rawText: String): TestSuiteGeneratedByLLM? {
         if (packageName.isBlank()) {
-            packageName = javaPackagePattern.find(rawText)?.groups?.get(1)?.value.orEmpty()
+            packageName = getPackageFromTestSuiteCode(rawText, Language.Java)
         }
 
         return JUnitTestSuiteParserStrategy.parseJUnitTestSuite(
