+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/kotlin/KotlinJUnitTestSuiteParser.kt
@@ -1,12 +1,13 @@
 package org.jetbrains.research.testspark.core.test.kotlin
 
 import org.jetbrains.research.testspark.core.data.JUnitVersion
+import org.jetbrains.research.testspark.core.generation.llm.getPackageFromTestSuiteCode
+import org.jetbrains.research.testspark.core.test.Language
 import org.jetbrains.research.testspark.core.test.TestBodyPrinter
 import org.jetbrains.research.testspark.core.test.TestSuiteParser
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.core.test.strategies.JUnitTestSuiteParserStrategy
 import org.jetbrains.research.testspark.core.utils.kotlinImportPattern
-import org.jetbrains.research.testspark.core.utils.kotlinPackagePattern
 
 class KotlinJUnitTestSuiteParser(
     private var packageName: String,
@@ -15,7 +16,7 @@ class KotlinJUnitTestSuiteParser(
 ) : TestSuiteParser {
     override fun parseTestSuite(rawText: String): TestSuiteGeneratedByLLM? {
         if (packageName.isBlank()) {
-            packageName = kotlinPackagePattern.find(rawText)?.groups?.get(1)?.value.orEmpty()
+            packageName = getPackageFromTestSuiteCode(rawText, Language.Kotlin)
         }
 
         return JUnitTestSuiteParserStrategy.parseJUnitTestSuite(
